WIP5

+++ projects/InjectDLL/game/pickups/tree.cpp
@@ -23,7 +23,7 @@ namespace {
     IL2CPP_INTERCEPT(, GetAbilityOnCondition, void, AssignAbility, (app::GetAbilityOnCondition * this_ptr)) {
         auto ability = this_ptr->fields.Ability->fields.Ability;
         if (is_tree(ability)) {
-            uber_states::UberState(UberStateGroup::Tree, ability).set(1);
+            uber_states::UberState(UberStateGroup::Tree, static_cast<int>(ability)).set(1);
             csharp_bridge::on_tree(ability);
             // Update PlayerDescriptorUberState
             uber_states::UberState(static_cast<UberStateGroup>(3440), 37811).apply();
